### 2020-07-09 #2457 @whedon
[PRE REVIEW]: CoReCon: an open, community-powered collection of Reionization constraints
**Submitting author:** @EGaraldi (<a href="http://orcid.org/0000-0002-6021-7020">Enrico Garaldi</a>)
**Repository:** <a href="https://github.com/EGaraldi/corecon" target ="_blank">https://github.com/EGaraldi/corecon</a>
**Version:** v0.3.1
