### 2020-07-09 #2457 @kyleniemeyer
Hi @EGaraldi, thanks for your submission to JOSS. As we are getting started, could you please revise your paper to include an explicit Statement of Need section, as described in [our documentation](https://joss.readthedocs.io/en/latest/submitting.html#what-should-my-paper-contain)? Also, it looks like some DOIs may be missing for some references. Thanks!
