### 2020-08-04 EGaraldi/corecon#6 @Stefan-Heimersheim
Suggestions for the documentation
Hello,

First let me say that I think this project will be very useful and the collection of constraints is a great idea. I have been taking a look at it for the [JOSS Review](https://github.com/openjournals/joss-reviews/issues/2495) and want to suggest a few improvements to the documentation here. (I'll put unrelated comments or bug reports in separate issues to keep everything clear.)

The table of contents on the [readthedocs](https://corecon.readthedocs.io/en/latest/index.html) page doesn't contain all the items (e.g. tutorial, the constraints, function docs) and when you click on one of the links (examples from before) the table of contents disappears. Would it be possible to include them all in the sidebar? That would make it considerably easier to navigate and find things? See e.g. [cobaya](https://cobaya.readthedocs.io/en/latest/input.html) or [this repo](https://gamma-astro-data-formats.readthedocs.io/en/latest/general/coordinates.html) for examples.

You have some nice plots illustrating the various constraints (e.g. [here](https://corecon.readthedocs.io/en/latest/datarst/ionized_fraction.html)). Could you add/link the source code for those plots as well? This would be helpful for contributors adding their constraints so they can update the plots, but also for users using the constraints.

Best,
Stefan
