### 2020-08-05 096b600
Assert that field in get() is valid
Throw an error if the user tries to obtain a field that does not exist,
e.g. with a typo like crc.get('To').
Otherwise the return None value will cause problems in the users code
later and is harder to debug.
