### 2020-08-05 8c1df42
Added data for Qin et al. 2020 xHI(z=10) constraint from CMB
