+++ corecon/data/ionized_fraction/Qin_et_al_2020.py
@@ -1,0 +1,34 @@
+dictionary_tag         = "Qin et al. 2020 (subm.)"
+
+reference              = "Qin, Poulin, Mesinger, Greig, Murray, Park, Jaehong; subm. (2020)"
+
+url                    = "https://arxiv.org/pdf/2006.16828.pdf"
+
+description            = \
+"""
+Ionized fraction for physical models (simulated with 21cmFAST) that satisfy the CMB optical depth constraint (Planck 2018) and also an x_HI(5.9)<0.06+0.05(1σ) constraint from QSO dark fraction. Note that this constraint is model dependent, for a simplistic tanh model it would be stronger.
+"""
+
+data_structure         = "" #grid or points
+
+extracted              = False #False if the original paper provides number, True if extracted from plots
+
+ndim                   = 1
+
+dimensions_descriptors = [ redshift ]
+
+axes                   = [ [10.] ]
+
+values                 = [ 0.849 ]
+
+err_up                 = [ nan ]
+
+err_down               = [ nan ]
+
+err_up2                = [ nan ]
+
+err_down2              = [ nan ]
+
+upper_lim              = [ False ]
+
+lower_lim              = [ True ]
