### 2020-08-05 ac8355c
Fixed Qin et al 2020 constraint on ionized fraction
