+++ corecon/data/ionized_fraction/Qin_et_al_2020.py
@@ -9,25 +9,25 @@ description            = \
 Ionized fraction for physical models (simulated with 21cmFAST) that satisfy the CMB optical depth constraint (Planck 2018) and also an x_HI(5.9)<0.06+0.05(1σ) constraint from QSO dark fraction. Note that this constraint is model dependent, for a simplistic tanh model it would be stronger.
 """
 
-data_structure         = "" #grid or points
+data_structure         = "points" #grid or points
 
 extracted              = False #False if the original paper provides number, True if extracted from plots
 
 ndim                   = 1
 
-dimensions_descriptors = [ redshift ]
+dimensions_descriptors = [ "redshift" ]
 
-axes                   = [ [10.] ]
+axes                   = [ 10 ]
 
-values                 = [ 0.849 ]
+values                 = [ 1-0.849 ]
 
-err_up                 = [ nan ]
+err_up                 = [ None ]
 
-err_down               = [ nan ]
+err_down               = [ None ]
 
-err_up2                = [ nan ]
+err_up2                = [ None ]
 
-err_down2              = [ nan ]
+err_down2              = [ None ]
 
 upper_lim              = [ False ]
 
