+++ corecon/data/ionized_fraction/Qin_et_al_2020.py
@@ -29,6 +29,6 @@ err_up2                = [ None ]
 
 err_down2              = [ None ]
 
-upper_lim              = [ False ]
+upper_lim              = [ True ]
 
-lower_lim              = [ True ]
+lower_lim              = [ False ]
