### 2020-08-05 EGaraldi/corecon#6 @Stefan-Heimersheim
Another small point: Would you mind adding a note to your docs on how to report issues or problems with the software and to seek support, fulfilling the requirements for the [community guidelines](https://joss.readthedocs.io/en/latest/review_criteria.html#community-guidelines)?
