### 2020-09-03 6c5b35e
DOCS: data sources now in alphabetical order
