+++ corecon/data/tau_cmb/Paoletti_et_al_2020.py
@@ -1,0 +1,34 @@
+dictionary_tag         = "Paoletti et al. 2020"
+
+reference              = "Paoletti D., Hazra D.K., Finelli F., Smoot G.F. 2020, JCAP submitted"
+
+url                    = "https://ui.adsabs.harvard.edu/abs/2020arXiv200512222P/abstract"
+
+description            = \
+"""
+Thomson scattering optical depth obtained from an updated analysis of Planck 2018 CMB data
+"""
+
+data_structure         = "points" #grid or points
+
+extracted              = False #False if the original paper provides number, True if extracted from plots
+
+ndim                   = 0
+
+dimensions_descriptors = []
+
+axes                   = []
+
+values                 = [0.0572]
+
+err_up                 = [0.0064]
+
+err_down               = [0.0075] 
+
+err_up2                = [None]
+
+err_down2              = [None]
+
+upper_lim              = [None]
+
+lower_lim              = [None]
