+++ docs/datarst/glf.rst
@@ -8,59 +8,11 @@ Galaxy luminosity function
 Data sources
 ^^^^^^^^^^^^
 
-|Ishigaki et al. 2018|
-
-.. |Ishigaki et al. 2018| raw:: html
-
-   <a href="https://iopscience.iop.org/article/10.3847/1538-4357/aaa544" target="_blank">Ishigaki et al. 2018</a>
-
 |Atek et al. 2018|
 
 .. |Atek et al. 2018| raw:: html
 
    <a href="https://academic.oup.com/mnras/article/479/4/5184/5050078" target="_blank">Atek et al. 2018</a>
-
-|Oesch et al. 2014|
-
-.. |Oesch et al. 2014| raw:: html
-
-   <a href="https://iopscience.iop.org/article/10.1088/0004-637X/786/2/108" target="_blank">Oesch et al. 2014</a>
-
-|Morishita et al. 2018|
-
-.. |Morishita et al. 2018| raw:: html
-
-   <a href="https://iopscience.iop.org/article/10.3847/1538-4357/aae68c" target="_blank">Morishita et al. 2018</a>
-
-|Bouwens et al. 2016|
-
-.. |Bouwens et al. 2016| raw:: html
-
-   <a href="https://iopscience.iop.org/article/10.3847/0004-637X/830/2/67" target="_blank">Bouwens et al. 2016</a>
-
-|Livermore et a. 2017|
-
-.. |Livermore et a. 2017| raw:: html
-
-   <a href="https://iopscience.iop.org/article/10.3847/1538-4357/835/2/113" target="_blank">Livermore et a. 2017</a>
-
-|McLeod et al. 2016|
-
-.. |McLeod et al. 2016| raw:: html
-
-   <a href="https://academic.oup.com/mnras/article/459/4/3812/2624050" target="_blank">McLeod et al. 2016</a>
-
-|Finkelstein et al. 2015|
-
-.. |Finkelstein et al. 2015| raw:: html
-
-   <a href="https://iopscience.iop.org/article/10.1088/0004-637X/810/1/71" target="_blank">Finkelstein et al. 2015</a>
-
-|Stefanon et al. 2019|
-
-.. |Stefanon et al. 2019| raw:: html
-
-   <a href="https://iopscience.iop.org/article/10.3847/1538-4357/ab3792" target="_blank">Stefanon et al. 2019</a>
 
 |Bouwens et al. 2015|
 
@@ -68,17 +20,23 @@ Data sources
 
    <a href="https://iopscience.iop.org/article/10.1088/0004-637X/803/1/34" target="_blank">Bouwens et al. 2015</a>
 
-|McLure et al. 2013|
+|Bouwens et al. 2016|
 
-.. |McLure et al. 2013| raw:: html
+.. |Bouwens et al. 2016| raw:: html
 
-   <a href="https://academic.oup.com/mnras/article/432/4/2696/2907730" target="_blank">McLure et al. 2013</a>
+   <a href="https://iopscience.iop.org/article/10.3847/0004-637X/830/2/67" target="_blank">Bouwens et al. 2016</a>
 
 |Bouwens et al. 2017|
 
 .. |Bouwens et al. 2017| raw:: html
 
    <a href="https://iopscience.iop.org/article/10.3847/1538-4357/aa70a4" target="_blank">Bouwens et al. 2017</a>
+
+|Bowler et al. 2015|
+
+.. |Bowler et al. 2015| raw:: html
+
+   <a href="https://academic.oup.com/mnras/article/452/2/1817/1068199" target="_blank">Bowler et al. 2015</a>
 
 |Bowler et al. 2020|
 
@@ -92,9 +50,51 @@ Data sources
 
[…]
