+++ docs/datarst/qlf.rst
@@ -8,17 +8,41 @@ Quasar luminosity function
 Data sources
 ^^^^^^^^^^^^
 
-|Kashikawa et al. 2015|
+|Giallongo et al. 2015|
 
-.. |Kashikawa et al. 2015| raw:: html
+.. |Giallongo et al. 2015| raw:: html
 
-   <a href="https://iopscience.iop.org/article/10.1088/0004-637X/798/1/28" target="_blank">Kashikawa et al. 2015</a>
+   <a href="https://www.aanda.org/articles/aa/abs/2015/06/aa25334-14/aa25334-14.html" target="_blank">Giallongo et al. 2015</a>
+
+|Giallongo et al. 2019|
+
+.. |Giallongo et al. 2019| raw:: html
+
+   <a href="https://iopscience.iop.org/article/10.3847/1538-4357/ab39e1" target="_blank">Giallongo et al. 2019</a>
 
 |Glikman et al. 2011|
 
 .. |Glikman et al. 2011| raw:: html
 
    <a href="https://iopscience.iop.org/article/10.1088/2041-8205/728/2/L26" target="_blank">Glikman et al. 2011</a>
+
+|Jiang et al. 2016|
+
+.. |Jiang et al. 2016| raw:: html
+
+   <a href="https://iopscience.iop.org/article/10.3847/1538-4357/833/2/222" target="_blank">Jiang et al. 2016</a>
+
+|Kashikawa et al. 2015|
+
+.. |Kashikawa et al. 2015| raw:: html
+
+   <a href="https://iopscience.iop.org/article/10.1088/0004-637X/798/1/28" target="_blank">Kashikawa et al. 2015</a>
+
+|Kulkarni et al. 2019|
+
+.. |Kulkarni et al. 2019| raw:: html
+
+   <a href="https://academic.oup.com/mnras/article/488/1/1035/5510422" target="_blank">Kulkarni et al. 2019</a>
 
 |McGreer et al. 2013|
 
@@ -32,39 +56,15 @@ Data sources
 
    <a href="https://iopscience.iop.org/article/10.1088/0004-637X/773/1/14" target="_blank">Ross et al. 2013</a>
 
-|Jiang et al. 2016|
-
-.. |Jiang et al. 2016| raw:: html
-
-   <a href="https://iopscience.iop.org/article/10.3847/1538-4357/833/2/222" target="_blank">Jiang et al. 2016</a>
-
-|Yang et al. 2016|
-
-.. |Yang et al. 2016| raw:: html
-
-   <a href="https://iopscience.iop.org/article/10.3847/0004-637X/829/1/33" target="_blank">Yang et al. 2016</a>
-
 |Willott et al. 2010|
 
 .. |Willott et al. 2010| raw:: html
 
    <a href="https://iopscience.iop.org/article/10.1088/0004-6256/139/3/906" target="_blank">Willott et al. 2010</a>
 
-|Giallongo et al. 2019|
+|Yang et al. 2016|
 
-.. |Giallongo et al. 2019| raw:: html
+.. |Yang et al. 2016| raw:: html
 
-   <a href="https://iopscience.iop.org/article/10.3847/1538-4357/ab39e1" target="_blank">Giallongo et al. 2019</a>
-
-|Kulkarni et al. 2019|
-
-.. |Kulkarni et al. 2019| raw:: html
-
-   <a href="https://academic.oup.com/mnras/article/488/1/1035/5510422" target="_blank">Kulkarni et al. 2019</a>
-
-|Giallongo et al. 2015|
-
-.. |Giallongo et al. 2015| raw:: html
-
[…]
