+++ docs/datarst/Lya_spike_galaxy_correlation.rst
@@ -8,15 +8,15 @@ Ly-alpha spike - galaxy correlation
 Data sources
 ^^^^^^^^^^^^
 
-|Meyer et al. 2019|
-
-.. |Meyer et al. 2019| raw:: html
-
-   <a href="https://academic.oup.com/mnras/article/483/1/19/5159477" target="_blank">Meyer et al. 2019</a>
-
 |Kakiichi et al. 2018|
 
 .. |Kakiichi et al. 2018| raw:: html
 
    <a href="https://academic.oup.com/mnras/article/479/1/43/4999925" target="_blank">Kakiichi et al. 2018</a>
+
+|Meyer et al. 2019|
+
+.. |Meyer et al. 2019| raw:: html
+
+   <a href="https://academic.oup.com/mnras/article/483/1/19/5159477" target="_blank">Meyer et al. 2019</a>
 

[…]
