### 2020-09-03 c75eaba
MODULE: now auto-generate 'neutral_fraction' field + small updates to utility functions and docs
