### 2020-11-07 EGaraldi/corecon#15 @iliantiliev
Referee report (Iliev)
The CoRECon code is a useful collection of current reionization constraints
from various collections of observational data. By its nature, it is intended
as an expandable framework for including additional data when this becomes
available.

The code was easy to install and to get started with. The update mechanism
works well (though there was no new version of the code for me to test on).

However, the code can certainly do with some usability improvements. The one
and only worked out example provided did not work as given in the online
manual. First, the example code gave an error (which was easily fixed by
commenting a line, but should nonetheless not happen). Second, while it did
produce a plot as it was supposed to, that plot does not look anything like
what it is expected to according to the manual. The test code uses completely
different data sets from the example plot and is over different redshift
interval. This is confusing, and not very useful as a test if the installation
is working as expected. Some more, detailed (and properly working)
examples are needed to get people started.

Apart from this, the various available functions are reasonably well (if
briefly) documented. Adding additional constrains is also reasonably
straightforward, and this is important given the rapidly evolving field.
Somewhat less clear to me is if the author plans to regularly update the
constraints. If this code gets picked up by the community these updates
could be done by the community, but I am not sure how much this could be
relied on. Without regular updates this code is quickly going to become
obsolete. Most successful codes tend to have at least a few dedicated
contributors.

One thing that was not quite clear to me is who are the intended users of
this code. The people working in ths area for long time likely have some
form of this already implemented, so might not be the target users. More
likely this could be used by younger, less experienced researchers, like
PhD students, or undergraduates. If that is the case, the code should
probably be a bit more user-friendly than it currently is, to ensure this
is picked up by a wider community, which in turn will make it more useful
if additional data is continually being added.
