### 2020-11-10 e27bf93
A few typo corrections, and changes to the bibliography (notice the change in the citation for numpy, in line with the request at https://numpy.org/citing-numpy/).
