+++ paper/paper.bib
@@ -26,20 +26,31 @@ archivePrefix = {arXiv},
   number={3},
   pages={90--95},
   year={2007},
+  doi={10.1109/MCSE.2007.55},
   publisher={IEEE}
 }
 
-
-
-@article{numpy,
-  title={The NumPy array: a structure for efficient numerical computation},
-  author={Walt, St{\'e}fan van der and Colbert, S Chris and Varoquaux, Gael},
-  journal={Computing in Science \& Engineering},
-  volume={13},
-  number={2},
-  pages={22--30},
-  year={2011},
-  publisher={IEEE}
+@article{         harris2020array,
+ title         = {Array programming with {NumPy}},
+ author        = {Charles R. Harris and K. Jarrod Millman and St{'{e}}fan J.
+                 van der Walt and Ralf Gommers and Pauli Virtanen and David
+                 Cournapeau and Eric Wieser and Julian Taylor and Sebastian
+                 Berg and Nathaniel J. Smith and Robert Kern and Matti Picus
+                 and Stephan Hoyer and Marten H. van Kerkwijk and Matthew
+                 Brett and Allan Haldane and Jaime Fern{'{a}}ndez del
+                 R{'{\i}}o and Mark Wiebe and Pearu Peterson and Pierre
+                 G{'{e}}rard-Marchant and Kevin Sheppard and Tyler Reddy and
+                 Warren Weckesser and Hameer Abbasi and Christoph Gohlke and
+                 Travis E. Oliphant},
+ year          = {2020},
+ month         = sep,
+ journal       = {Nature},
+ volume        = {585},
+ number        = {7825},
+ pages         = {357--362},
+ doi           = {10.1038/s41586-020-2649-2},
+ publisher     = {Springer Science and Business Media {LLC}},
+ url           = {https://doi.org/10.1038/s41586-020-2649-2}
 }
 
 @dataset{qsolist,
