### 2020-11-10 EGaraldi/corecon#17 @eloisabentivegna
Reviewer comments, requests, and recommendation.
Dear @EGaraldi, thanks for your [submission](https://github.com/openjournals/joss-reviews/issues/2495)!

The package serves as a useful literature consolidator, as well as a proposal to standardize constraints from a cosmological era where few efforts already exist. It should be welcome both to this community and even more to researchers in neighboring fields (such as structure formation, or nuclear astrophysics) who may need this information but not have the background to digest loads of articles efficiently.

The package is functional and reasonably illustrated in the docs, although I second one of the other reviewers who pointed out that the tutorial does not work out of the box due to typos. These should be fixed before the paper can be formally accepted. The dependencies are also not clearly spelled out, although the package can be installed with `pip`, so the end user does not need to be aware of them (nonetheless, it would be good to state explicitly which versions of Python and the main packages are required). At the moment, there is also no mention of whether and how contributed data will be reviewed or validated before they are assimilated — if there are plans on how to handle incoming data, it would be nice to describe them in this paper.

One final question (which may not require any changes as much as a clarification) is whether you have considered including @benzene190 in the author list. Whilst their contributions are distinct from the core functionality of the package, they do add an interesting capability and might be considered worth recognizing with authorship. This is debatable: I am not saying @benzene190 needs to be an author, just asking whether you could clarify your decision.

Other than this, I have made a pull request with some corrections/suggestions for the main text. Once these are taken care of, I recommend acceptance. Very useful and welcome package!
