### 2020-11-10 #2495 @eloisabentivegna
I have just noticed that @iliantiliev has posted a reviewer report as an [issue](https://github.com/EGaraldi/corecon/issues/15) on the package's repository. I will touch base with him via email to see whether he wants to complete the checklist as well.
